requirements.txt
fastapi
uvicorn
PyMuPDF
transformers
torch
pandas
seqeval
presidio-image-redactor
presidio-analyzer
pillow 
tf-keras
tesseract